requirements.txt
dnspython
netaddr
requests
urllib3
